# Optimization: initialize vectors with capacity