@Geal 2019-02-17
yep, this definitely moves the needle. The way id o it for `many1` in nom, is that I apply the parser once, then if it was successful I allocate the `Vec`.